+++ guiyu-parent/guiyu-callCenter-parent/guiyu-callCenter-calloutserver/guiyu-callCenter-calloutserver-web/src/main/java/com/guiji/calloutserver/service/CallOutRecordService.java
@@ -2,6 +2,8 @@ package com.guiji.calloutserver.service;
 
 import com.guiji.callcenter.dao.entity.CallOutRecord;
 
+import java.math.BigInteger;
+
 /**
  * @Auther: 魏驰
  * @Date: 2018/11/4 16:56
@@ -9,7 +11,7 @@ import com.guiji.callcenter.dao.entity.CallOutRecord;
  * @Description:
  */
 public interface CallOutRecordService {
-    CallOutRecord findByCallId(Long callId);
+    CallOutRecord findByCallId(BigInteger callId);
     void save(CallOutRecord callOutRecord);
 
     void update(CallOutRecord callOutRecord);
